+++ src/modules/person-kontext/api/personenkontext.uc.spec.ts
@@ -5,19 +5,19 @@ import { DoFactory, MapperTestModule } from '../../../../test/utils/index.js';
 import { EntityCouldNotBeCreated } from '../../../shared/error/entity-could-not-be-created.error.js';
 import { EntityNotFoundError } from '../../../shared/error/entity-not-found.error.js';
 import { Paged } from '../../../shared/paging/paged.js';
-import { PersonDo } from '../domain/person.do.js';
-import { PersonService } from '../domain/person.service.js';
-import { PersonenkontextDo } from '../domain/personenkontext.do.js';
-import { Personenstatus, Rolle, SichtfreigabeType } from '../domain/personenkontext.enums.js';
-import { PersonenkontextService } from '../domain/personenkontext.service.js';
+import { PersonDo } from '../../person/domain/person.do.js';
+import { PersonService } from '../../person/domain/person.service.js';
+import { PersonenkontextDo } from '../../person/domain/personenkontext.do.js';
+import { Personenstatus, Rolle, SichtfreigabeType } from '../../person/domain/personenkontext.enums.js';
+import { PersonenkontextService } from '../../person/domain/personenkontext.service.js';
 import { CreatePersonenkontextDto } from './create-personenkontext.dto.js';
 import { CreatedPersonenkontextDto } from './created-personenkontext.dto.js';
 import { FindPersonenkontextByIdDto } from './find-personenkontext-by-id.dto.js';
 import { FindPersonenkontextDto } from './find-personenkontext.dto.js';
-import { PersonApiMapperProfile } from './person-api.mapper.profile.js';
+import { PersonApiMapperProfile } from '../../person/api/person-api.mapper.profile.js';
 import { PersonenkontextDto } from './personenkontext.dto.js';
 import { PersonenkontextUc } from './personenkontext.uc.js';
-import { PersonendatensatzDto } from './personendatensatz.dto.js';
+import { PersonendatensatzDto } from '../../person/api/personendatensatz.dto.js';
 import { SchulConnexError } from '../../../shared/error/schul-connex.error.js';
 import { DeletePersonenkontextDto } from './delete-personkontext.dto.js';
