+++ src/modules/person-kontext/api/personenkontext.uc.ts
@@ -2,16 +2,16 @@ import { Mapper } from '@automapper/core';
 import { getMapperToken } from '@automapper/nestjs';
 import { Inject, Injectable } from '@nestjs/common';
 import { Paged } from '../../../shared/paging/paged.js';
-import { PersonDo } from '../domain/person.do.js';
-import { PersonService } from '../domain/person.service.js';
-import { PersonenkontextDo } from '../domain/personenkontext.do.js';
-import { PersonenkontextService } from '../domain/personenkontext.service.js';
+import { PersonDo } from '../../person/domain/person.do.js';
+import { PersonService } from '../../person/domain/person.service.js';
+import { PersonenkontextDo } from '../../person/domain/personenkontext.do.js';
+import { PersonenkontextService } from '../../person/domain/personenkontext.service.js';
 import { CreatePersonenkontextDto } from './create-personenkontext.dto.js';
 import { CreatedPersonenkontextDto } from './created-personenkontext.dto.js';
 import { FindPersonenkontextByIdDto } from './find-personenkontext-by-id.dto.js';
 import { FindPersonenkontextDto } from './find-personenkontext.dto.js';
-import { PersonDto } from './person.dto.js';
-import { PersonendatensatzDto } from './personendatensatz.dto.js';
+import { PersonDto } from '../../person/api/person.dto.js';
+import { PersonendatensatzDto } from '../../person/api/personendatensatz.dto.js';
 import { PersonenkontextDto } from './personenkontext.dto.js';
 import { UpdatePersonenkontextDto } from './update-personenkontext.dto.js';
 import { DomainError } from '../../../shared/error/domain.error.js';
