+++ src/modules/person-kontext/api/personenkontextdatensatz.response.ts
@@ -1,6 +1,6 @@
 import { AutoMap } from '@automapper/classes';
 import { ApiProperty } from '@nestjs/swagger';
-import { PersonIdResponse } from './person-id.response.js';
+import { PersonIdResponse } from '../../person/api/person-id.response.js';
 import { PersonenkontextResponse } from './personenkontext.response.js';
 
 export class PersonenkontextdatensatzResponse {
