+++ src/modules/person-kontext/api/update-personenkontext.dto.ts
@@ -1,5 +1,5 @@
 import { AutoMap } from '@automapper/classes';
-import { Jahrgangsstufe, Personenstatus } from '../domain/personenkontext.enums.js';
+import { Jahrgangsstufe, Personenstatus } from '../../person/domain/personenkontext.enums.js';
 
 export class UpdatePersonenkontextDto {
     @AutoMap()
