+++ src/modules/person/api/person-api.mapper.profile.spec.ts
@@ -9,12 +9,12 @@ import { PersonenkontextDo } from '../domain/personenkontext.do.js';
 import { Jahrgangsstufe, Personenstatus, Rolle, SichtfreigabeType } from '../domain/personenkontext.enums.js';
 import { CreatePersonBodyParams } from './create-person.body.params.js';
 import { CreatePersonDto } from './create-person.dto.js';
-import { CreatePersonenkontextBodyParams } from './create-personenkontext.body.params.js';
-import { CreatePersonenkontextDto } from './create-personenkontext.dto.js';
-import { CreatedPersonenkontextDto } from './created-personenkontext.dto.js';
-import { FindPersonenkontextByIdDto } from './find-personenkontext-by-id.dto.js';
-import { FindPersonenkontextByIdParams } from './find-personenkontext-by-id.params.js';
-import { FindPersonenkontextDto } from './find-personenkontext.dto.js';
+import { CreatePersonenkontextBodyParams } from '../../person-kontext/api/create-personenkontext.body.params.js';
+import { CreatePersonenkontextDto } from '../../person-kontext/api/create-personenkontext.dto.js';
+import { CreatedPersonenkontextDto } from '../../person-kontext/api/created-personenkontext.dto.js';
+import { FindPersonenkontextByIdDto } from '../../person-kontext/api/find-personenkontext-by-id.dto.js';
+import { FindPersonenkontextByIdParams } from '../../person-kontext/api/find-personenkontext-by-id.params.js';
+import { FindPersonenkontextDto } from '../../person-kontext/api/find-personenkontext.dto.js';
 import { LoeschungDto } from './loeschung.dto.js';
 import { LoeschungResponse } from './loeschung.response.js';
 import { PersonApiMapperProfile } from './person-api.mapper.profile.js';
@@ -26,12 +26,12 @@ import { PersonDto } from './person.dto.js';
 import { PersonResponse } from './person.response.js';
 import { PersonendatensatzDto } from './personendatensatz.dto.js';
 import { PersonendatensatzResponse } from './personendatensatz.response.js';
-import { PersonenkontextQueryParams } from './personenkontext-query.params.js';
-import { PersonenkontextDto } from './personenkontext.dto.js';
-import { PersonenkontextResponse } from './personenkontext.response.js';
-import { PersonenkontextdatensatzResponse } from './personenkontextdatensatz.response.js';
-import { UpdatePersonenkontextDto } from './update-personenkontext.dto.js';
-import { UpdatePersonenkontextBodyParams } from './update-personenkontext.body.params.js';
+import { PersonenkontextQueryParams } from '../../person-kontext/api/personenkontext-query.params.js';
+import { PersonenkontextDto } from '../../person-kontext/api/personenkontext.dto.js';
+import { PersonenkontextResponse } from '../../person-kontext/api/personenkontext.response.js';
+import { PersonenkontextdatensatzResponse } from '../../person-kontext/api/personenkontextdatensatz.response.js';
+import { UpdatePersonenkontextDto } from '../../person-kontext/api/update-personenkontext.dto.js';
+import { UpdatePersonenkontextBodyParams } from '../../person-kontext/api/update-personenkontext.body.params.js';
 import { UpdatePersonBodyParams } from './update-person.body.params.js';
 import { UpdatePersonDto } from './update-person.dto.js';
