+++ src/modules/person/api/person-api.mapper.profile.ts
@@ -7,14 +7,14 @@ import { PersonDo } from '../domain/person.do.js';
 import { PersonenkontextDo } from '../domain/personenkontext.do.js';
 import { CreatePersonBodyParams } from './create-person.body.params.js';
 import { CreatePersonDto } from './create-person.dto.js';
-import { CreatePersonenkontextBodyParams } from './create-personenkontext.body.params.js';
-import { CreatePersonenkontextDto } from './create-personenkontext.dto.js';
-import { CreatedPersonenkontextOrganisationDto } from './created-personenkontext-organisation.dto.js';
-import { CreatedPersonenkontextDto } from './created-personenkontext.dto.js';
+import { CreatePersonenkontextBodyParams } from '../../person-kontext/api/create-personenkontext.body.params.js';
+import { CreatePersonenkontextDto } from '../../person-kontext/api/create-personenkontext.dto.js';
+import { CreatedPersonenkontextOrganisationDto } from '../../person-kontext/api/created-personenkontext-organisation.dto.js';
+import { CreatedPersonenkontextDto } from '../../person-kontext/api/created-personenkontext.dto.js';
 import { FindPersonendatensatzDto } from './find-personendatensatz.dto.js';
-import { FindPersonenkontextByIdDto } from './find-personenkontext-by-id.dto.js';
-import { FindPersonenkontextByIdParams } from './find-personenkontext-by-id.params.js';
-import { FindPersonenkontextDto } from './find-personenkontext.dto.js';
+import { FindPersonenkontextByIdDto } from '../../person-kontext/api/find-personenkontext-by-id.dto.js';
+import { FindPersonenkontextByIdParams } from '../../person-kontext/api/find-personenkontext-by-id.params.js';
+import { FindPersonenkontextDto } from '../../person-kontext/api/find-personenkontext.dto.js';
 import { LoeschungDto } from './loeschung.dto.js';
 import { LoeschungResponse } from './loeschung.response.js';
 import { PersonBirthParams } from './person-birth.params.js';
@@ -27,16 +27,16 @@ import { PersonResponse } from './person.response.js';
 import { PersonenQueryParams } from './personen-query.param.js';
 import { PersonendatensatzDto } from './personendatensatz.dto.js';
 import { PersonendatensatzResponse } from './personendatensatz.response.js';
-import { PersonenkontextQueryParams } from './personenkontext-query.params.js';
-import { PersonenkontextDto } from './personenkontext.dto.js';
-import { PersonenkontextResponse } from './personenkontext.response.js';
-import { PersonenkontextdatensatzResponse } from './personenkontextdatensatz.response.js';
-import { UpdatePersonenkontextBodyParams } from './update-personenkontext.body.params.js';
-import { UpdatePersonenkontextDto } from './update-personenkontext.dto.js';
+import { PersonenkontextQueryParams } from '../../person-kontext/api/personenkontext-query.params.js';
+import { PersonenkontextDto } from '../../person-kontext/api/personenkontext.dto.js';
+import { PersonenkontextResponse } from '../../person-kontext/api/personenkontext.response.js';
+import { PersonenkontextdatensatzResponse } from '../../person-kontext/api/personenkontextdatensatz.response.js';
+import { UpdatePersonenkontextBodyParams } from '../../person-kontext/api/update-personenkontext.body.params.js';
+import { UpdatePersonenkontextDto } from '../../person-kontext/api/update-personenkontext.dto.js';
 import { UpdatePersonBodyParams } from './update-person.body.params.js';
 import { UpdatePersonDto } from './update-person.dto.js';
 import { DeleteRevisionBodyParams } from './delete-revision.body.params.js';
-import { DeletePersonenkontextDto } from './delete-personkontext.dto.js';
+import { DeletePersonenkontextDto } from '../../person-kontext/api/delete-personkontext.dto.js';
 
 @Injectable()
 export class PersonApiMapperProfile extends AutomapperProfile {
