+++ src/modules/person/api/person.controller.ts
@@ -35,20 +35,20 @@ import { ResultInterceptor } from '../../../shared/util/result-interceptor.js';
 import { PersonUc } from '../api/person.uc.js';
 import { CreatePersonBodyParams } from './create-person.body.params.js';
 import { CreatePersonDto } from './create-person.dto.js';
-import { CreatePersonenkontextBodyParams } from './create-personenkontext.body.params.js';
-import { CreatePersonenkontextDto } from './create-personenkontext.dto.js';
-import { CreatedPersonenkontextDto } from './created-personenkontext.dto.js';
+import { CreatePersonenkontextBodyParams } from '../../person-kontext/api/create-personenkontext.body.params.js';
+import { CreatePersonenkontextDto } from '../../person-kontext/api/create-personenkontext.dto.js';
+import { CreatedPersonenkontextDto } from '../../person-kontext/api/created-personenkontext.dto.js';
 import { FindPersonendatensatzDto } from './find-personendatensatz.dto.js';
-import { FindPersonenkontextDto } from './find-personenkontext.dto.js';
+import { FindPersonenkontextDto } from '../../person-kontext/api/find-personenkontext.dto.js';
 import { PersonByIdParams } from './person-by-id.param.js';
 import { PersonDto } from './person.dto.js';
 import { PersonenQueryParams } from './personen-query.param.js';
 import { PersonendatensatzDto } from './personendatensatz.dto.js';
 import { PersonendatensatzResponse } from './personendatensatz.response.js';
-import { PersonenkontextQueryParams } from './personenkontext-query.params.js';
-import { PersonenkontextDto } from './personenkontext.dto.js';
-import { PersonenkontextResponse } from './personenkontext.response.js';
-import { PersonenkontextUc } from './personenkontext.uc.js';
+import { PersonenkontextQueryParams } from '../../person-kontext/api/personenkontext-query.params.js';
+import { PersonenkontextDto } from '../../person-kontext/api/personenkontext.dto.js';
+import { PersonenkontextResponse } from '../../person-kontext/api/personenkontext.response.js';
+import { PersonenkontextUc } from '../../person-kontext/api/personenkontext.uc.js';
 import { UpdatePersonBodyParams } from './update-person.body.params.js';
 import { UpdatePersonDto } from './update-person.dto.js';
