+++ src/modules/person/api/person.uc.ts
@@ -13,10 +13,10 @@ import { SichtfreigabeType } from '../domain/personenkontext.enums.js';
 import { PersonenkontextService } from '../domain/personenkontext.service.js';
 import { CreatePersonDto } from './create-person.dto.js';
 import { FindPersonendatensatzDto } from './find-personendatensatz.dto.js';
-import { FindPersonenkontextDto } from './find-personenkontext.dto.js';
+import { FindPersonenkontextDto } from '../../person-kontext/api/find-personenkontext.dto.js';
 import { PersonDto } from './person.dto.js';
 import { PersonendatensatzDto } from './personendatensatz.dto.js';
-import { PersonenkontextDto } from './personenkontext.dto.js';
+import { PersonenkontextDto } from '../../person-kontext/api/personenkontext.dto.js';
 import { UserRepository } from '../../user/user.repository.js';
 import { User } from '../../user/user.js';
 import { ClassLogger } from '../../../core/logging/class-logger.js';
