+++ src/modules/person/api/personendatensatz.dto.ts
@@ -1,6 +1,6 @@
 import { AutoMap } from '@automapper/classes';
 import { PersonDto } from './person.dto.js';
-import { PersonenkontextDto } from './personenkontext.dto.js';
+import { PersonenkontextDto } from '../../person-kontext/api/personenkontext.dto.js';
 
 export class PersonendatensatzDto {
     @AutoMap(() => PersonDto)
