+++ src/modules/person/api/personendatensatz.response.ts
@@ -1,7 +1,7 @@
 import { AutoMap } from '@automapper/classes';
 import { ApiProperty } from '@nestjs/swagger';
 import { PersonResponse } from './person.response.js';
-import { PersonenkontextResponse } from './personenkontext.response.js';
+import { PersonenkontextResponse } from '../../person-kontext/api/personenkontext.response.js';
 
 export class PersonendatensatzResponse {
     @AutoMap(() => PersonResponse)
