+++ src/modules/person/person-api.module.ts
@@ -5,8 +5,8 @@ import { PersonController } from './api/person.controller.js';
 import { PersonUc } from './api/person.uc.js';
 import { PersonModule } from './person.module.js';
 import { KeycloakAdministrationModule } from '../keycloak-administration/keycloak-administration.module.js';
-import { PersonenkontextUc } from './api/personenkontext.uc.js';
-import { PersonenkontextController } from './api/personenkontext.controller.js';
+import { PersonenkontextUc } from '../person-kontext/api/personenkontext.uc.js';
+import { PersonenkontextController } from '../person-kontext/api/personenkontext.controller.js';
 import { UserModule } from '../user/user.module.js';
 import { PersonFrontendController } from './api/person.frontend.controller.js';
